Translate Java longs to double in V8ObjectUtils
Java long (64bit) was being stored as an int (32bit) which resulted
in numbers being truncated. Instead, store longs as doubles.

This fixes #208

+++ src/main/java/com/eclipsesource/v8/utils/V8ObjectUtils.java
@@ -455,7 +455,7 @@ public class V8ObjectUtils {
         } else if (value instanceof Integer) {
             result.add(key, (Integer) value);
         } else if (value instanceof Long) {
-            result.add(key, (int) (long) (Long) value);
+            result.add(key, (Long) value);
         } else if (value instanceof Double) {
             result.add(key, (Double) value);
         } else if (value instanceof Float) {
